requirements.txt
chainer
flake8
flask
jupyter
lru-dict
numpy
pandas
pytest
seaborn
